Moved tokenautocomplete to People Picker package
