+++ fluentui_peoplepicker/src/main/java/com/microsoft/fluentui/tokenautocomplete/FilteredArrayAdapter.java
@@ -1,4 +1,4 @@
-package com.microsoft.fluentui.util.tokenautocomplete;
+package com.microsoft.fluentui.tokenautocomplete;
 
 import android.content.Context;
 import android.widget.ArrayAdapter;
